cleanup

+++ src/screens/nc_lyrics.c
@@ -951,7 +951,7 @@ native_lyrics_report_save_error(NcmBuffer *filename, NcmError *error) {
     }
     args[0] = ncm_string_format_arg_string(filename->data, filename->len);
     args[1] = ncm_string_format_arg_cstring(message);
-    ncm_statusbar_format((int32)Config.message_delay_time,
+    ncm_statusbar_format(Config.message_delay_time,
                          STRLIT_ARGS("Couldn't save lyrics as \"%1%\": %2%"),
                          args, NCM_ARRAY_LEN(args));
     return;
@@ -968,7 +968,7 @@ native_lyrics_report_unlink_error(NcmBuffer *filename, NcmError *error) {
     }
     args[0] = ncm_string_format_arg_string(filename->data, filename->len);
     args[1] = ncm_string_format_arg_cstring(message);
-    ncm_statusbar_format((int32)Config.message_delay_time,
+    ncm_statusbar_format(Config.message_delay_time,
                          STRLIT_ARGS("Couldn't remove \"%1%\": %2%"),
                          args, NCM_ARRAY_LEN(args));
     return;
